+++ lib/Constants.js
@@ -684,9 +684,6 @@ const SortDirection = {
 };
 
 const SortKeys = {
-  Downloads: {
-
-  },
   Requests: {
     ASSIGNED_TO:
       r => (r.studyRequest.assignedTo === null
